refactor(canopy): extract shared query logic to engine server utility
Create fetchCanopyDirectory() in engine to eliminate ~160 lines of
duplicated query logic between API endpoint and page server load.

Changes:
- engine/src/lib/server/canopy-directory.ts: New shared utility
- engine/src/lib/server/index.ts: Export new utility
- landing/api/canopy/+server.ts: Use shared utility (~70 lines removed)
- landing/canopy/+page.server.ts: Use shared utility (~80 lines removed)

Result: Single source of truth for canopy directory queries,
following engine-first pattern consistently.

Addresses final code review feedback on PR #1017
